+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - added RSS source type
 
+### Changed
+- use staging auth by default
+
 ## [0.2.0] - 13.10.2019
 ### Added
 - added proposed event accept functionality